Update code menu manage

+++ dakmark/app/Http/Controllers/Admin/NavigatorController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use App\Models\Navigator;
+use App\Models\Seo;
 use DB;
 
 class NavigatorController extends Controller
@@ -16,7 +17,7 @@ class NavigatorController extends Controller
     }
     public function addNavigator(){
         $navigators =  Navigator::where("type",1)->orderBy('sort_order', 'asc')->get();  // lấy danh sách menu được tạo trực tiếp
-        $system_cats = Navigator::where("type",1)->orWhere('type',3)->get(); // lấy tất cả danh mục sản phẩm và danh mục blogs
+        $system_cats = Seo::where("type","PCAT")->orWhere("type","BCAT")->get(); // lấy tất cả danh mục sản phẩm và danh mục blogs
         return view('admin.navigator.add_navigator')->with(["navigators" => $navigators, "system_cats" => $system_cats]);
     } 
     public function insertNavigator(Request $request){
@@ -24,93 +25,62 @@ class NavigatorController extends Controller
                                   'slug' => 'required|unique:seo',
                                   'seo_title' => 'required'
                                 ]);
-        $max_id = $this->get_max_id('product_cat');
-        $system_id = 'PCAT'.$max_id;
-        $icon = '' ;
-        $icon_file = $request->file('icon');
-        if($icon_file != NULL){
-            $path = './public/assets/img/product_cat/';
-            if(!is_dir($path)){
-                mkdir($path, 0777, true);
-            }   
-            $icon = $this->upload_file($request->name, $icon_file, $path);
-        }
+        $max_id = $this->get_max_id('navigator');
+        $system_id = 'NAV'.$max_id;
 
-        // Insert data vào bảng product_cat
-        $productCat = new ProductCat;
-        $productCat->name = $request->name;
-        $productCat->en_name = $request->en_name;
-        $productCat->system_id = $system_id;
-        $productCat->parent_id = $request->parent_id;
-        $productCat->icon = $icon;
-        $productCat->sort_order = $request->sort_order;
-        $productCat->is_show = $request->is_show;
-        $productCat->is_show_nav = $request->is_show_nav;
-        $productCat->save();
+        // Insert data vào bảng navigator
+        $navigator = new Navigator;
+        $navigator->system_id = $system_id;
+        $navigator->name = $request->name;
+        $navigator->en_name = $request->en_name;
+        $navigator->parent_id = $request->parent_id;
+        $navigator->sort_order = $request->sort_order;
+        $navigator->is_show = $request->is_show;
+        $navigator->type = 1; // menu được tạo trực tiếp : type = 1
+        $navigator->save();
 
         // Insert data vào bảng seo
         $seo = new Seo;
         $seo->system_id = $system_id;
+        $seo->name = $request->name;
         $seo->slug = $request->slug;
         $seo->seo_title = $request->seo_title;
         $seo->en_seo_title = $request->en_seo_title;
         $seo->keyword = $request->keyword;
         $seo->description = $request->description;
-        $seo->type = 1;  // Danh mục sản phẩm : type = 1
+        $seo->type = "NAVIGATOR";  
         $seo->save();
 
-        // Insert data vào bảng navigator (menu)
-        if($request->is_show_nav==1){
-            $navigator = new Navigator;
-            $navigator->system_id = $system_id;
-            $navigator->name = $request->name;
-            $navigator->type = 2;   // Menu được tạo gián tiếp (qua danh mục sản phẩm hoặc danh mục bào viết ) : type = 2
-            $navigator->save();
-        }
-
-        session()->flash('success_message', "Thêm danh mục thành công !");
-        return redirect()->route('admin.product-cat');
+        session()->flash('success_message', "Thêm menu thành công !");
+        return redirect()->route('admin.navigator');
     }  
     public function editNavigator($nav_id)
     {
-        $productCatList = ProductCat::where('parent_id',0)->orderBy('sort_order', 'asc')->get();
-        $productCatDetail = ProductCat::where('id',$cat_id)->first();
-        $productCatSeo = Seo::where('system_id',$productCatDetail->system_id)->first();
-        return view('admin.products.edit_product_cat')->with(["productCatList" => $productCatList, 
-                                                              "productCatDetail" => $productCatDetail, 
-                                                              "productCatSeo" => $productCatSeo]); 
+        $navigatorList = Navigator::where('parent_id',0)->orderBy('sort_order', 'asc')->get();
+        $navigatorDetail = Navigator::where('id',$nav_id)->first();
+        $navigatorSeo = Seo::where('system_id',$navigatorDetail->system_id)->first();
+        return view('admin.navigator.edit_navigator')->with(["navigatorList" => $navigatorList, 
+                                                              "navigatorDetail" => $navigatorDetail, 
+                                                              "navigatorSeo" => $navigatorSeo]); 
     }
     public function updateNavigator($nav_id, Request $request)
     {
-        $productCat = ProductCat::find($cat_id);
+        $navigator = Navigator::find($nav_id);
         $this->validate($request,['name' => 'required',
-                                  'slug' => 'required',
+                                  'slug' => 'required',  // thiếu check_exist
                                   'seo_title' => 'required'
                                 ]);
-        $icon = '' ;
-        $icon_file = $request->file('icon');
-        if($icon_file != NULL){
-            $path = './public/assets/img/product_cat/';
-            if(!is_dir($path)){
-                mkdir($path, 0777, true);
-            }   
-            $icon = $this->upload_file($request->name, $icon_file, $path);
-        }
 
-        
-
-        // Update data bảng product_cat
-        $productCat->name = $request->name;
-        $productCat->en_name = $request->en_name;
-        $productCat->parent_id = $request->parent_id;
-        $productCat->icon = $icon;
-        $productCat->sort_order = $request->sort_order;
-        $productCat->is_show = $request->is_show;
-        $productCat->is_show_nav = $request->is_show_nav;
-        $productCat->save();
+        // Update data bảng navigator
+        $navigator->name = $request->name;
+        $navigator->en_name = $request->en_name;
+        $navigator->parent_id = $request->parent_id;
+        $navigator->sort_order = $request->sort_order;
+        $navigator->is_show = $request->is_show;
+        $navigator->save();
 
         // Update data bảng seo
-        $seo = Seo::where('system_id',$productCat->system_id)->first();
+        $seo = Seo::where('system_id',$navigator->system_id)->first();
         $seo->slug = $request->slug;
         $seo->seo_title = $request->seo_title;
         $seo->en_seo_title = $request->en_seo_title;
@@ -118,40 +88,19 @@ class NavigatorController extends Controller
         $seo->description = $request->description;
         $seo->save();
 
-        if($productCat->is_show_nav != $request->is_show_nav){
-            if($request->is_show_nav==1){  // Insert data vào bảng navigator (menu)
-                $navigator = new Navigator;
-                $navigator->system_id = $productCat->system_id;
-                $navigator->name = $request->name;
-                $navigator->type = 2;   // Menu được tạo gián tiếp (qua danh mục sản phẩm hoặc danh mục bào viết ) : type = 2
-                $navigator->save();
-            }
-            else{  // Xóa menu
-                $navigator = Navigator::where('system_id',$productCat->system_id)->first();
-                $navigator->delete();
-            }
-        }
-
         session()->flash('success_message', "Cập nhật thành công !");
-        return redirect()->route('admin.product-cat'); 
+        return redirect()->route('admin.navigator'); 
     }
-    public function deleteNavigator($cat_id){
-        $productCat = ProductCat::find($cat_id);
-        if($productCat->icon != ''){
-            $icon_file = './public/assets/img/product_cat/'.$productCat->icon;
-            if(file_exists($icon_file))
-                unlink($icon_file);
+    public function deleteNavigator($nav_id){
+        $navigator = Navigator::find($nav_id);
+        if($navigator->type == 1){  // nếu menu được tạo trực tiếp thì xóa record seo tương ứng
+            $seo = Seo::where('system_id',$navigator->system_id)->first();
+            $seo->delete();
         }
-        if($productCat->is_show_nav==1){
-            $navigator = Navigator::where('system_id',$productCat->system_id)->first();
-            $navigator->delete();
-        }
-        $seo = Seo::where('system_id',$productCat->system_id)->first();
-        $seo->delete();
-        $productCat->delete();
+        $navigator->delete();
 
-        session()->flash('success_message', "Xóa danh mục thành công !");
-        return redirect()->route('admin.product-cat'); 
+        session()->flash('success_message', "Xóa menu thành công !");
+        return redirect()->route('admin.navigator'); 
     }
 
     // Hàm ajax lấy slug theo tên
@@ -188,23 +137,20 @@ class NavigatorController extends Controller
             return 1;
     }
 
-    // Upload 1 file
-    function upload_file($object_name, $file, $path){
-        $ext = $file->getClientOriginalExtension();
-        $org_name = url_format($object_name);   // Định dạng lại phần tên gốc (remove utf8)
-        $file_name = $org_name.".".$ext;
-        $i=0;
-        $check_file = file_exists($path.$file_name); // Kiểm tra file đã tồn tại trong thư mục hay chưa
-        while ($check_file){  // Nếu đã tồn tại thì thêm số đếm vào sau tên file
-            $i++;
-            $tmp_org_name = $org_name.'-'.$i;
-            $file_name = $tmp_org_name.".".$ext;
-            $check_file = file_exists($path.$file_name);
+    // Lấy thông tin danh mục hệ thống
+    function get_system_cat($system_id){
+        $data = array();
+        $system_cat = Seo::where('system_id',$system_id)->first();
+        if($system_cat != NULL){
+            $data = array('name' => $system_cat->name,
+                      'en_name' => $system_cat->en_name,
+                      'slug' => $system_cat->slug,
+                      'seo_title' => $system_cat->seo_title,
+                      'en_seo_title' => $system_cat->en_seo_title,
+                      'keyword' => $system_cat->keyword,
+                      'description' => $system_cat->description);
         }
-      
-        if($file->move($path,$file_name))
-            return $file_name;
-        else
-            return '';
+
+        echo json_encode($data);  
     }
 }
